Aded Makefiles to all subfolder in SciDac2007.

+++ SciDac2007/glob.cpp
@@ -14,7 +14,7 @@ std::vector<std::string> glob(std::string pattern)
   if (glob(pattern.c_str(), 0, 0, &pglob) != 0)
     v.push_back("?");
   else
-    for (unsigned int i = 0; i < pglob.gl_pathc; i++)
+    for (unsigned long i = 0; i < pglob.gl_pathc; i++)
       v.push_back(std::string(pglob.gl_pathv[i]));
   globfree(&pglob);
   return v;
@@ -23,7 +23,7 @@ std::vector<std::string> glob(std::string pattern)
 int main()
 {
   std::vector<std::string> v = glob("*.vtk");
-  for (int i = 0; i < v.size(); i++)
+  for (size_t i = 0; i < v.size(); i++)
     std::cout << v[i] << std::endl;
   return 0;
 }
